### 2023-01-30 182b36f
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/charliermarsh/ruff-pre-commit: v0.0.230 → v0.0.237](https://github.com/charliermarsh/ruff-pre-commit/compare/v0.0.230...v0.0.237)
