+++ .pre-commit-config.yaml
@@ -6,7 +6,7 @@ repos:
       - id: black
         language_version: python3
   - repo: https://github.com/charliermarsh/ruff-pre-commit
-    rev: "v0.0.230"
+    rev: "v0.0.237"
     hooks:
       - id: ruff
 
