### 2023-02-06 d34a976
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/psf/black: 22.12.0 → 23.1.0](https://github.com/psf/black/compare/22.12.0...23.1.0)
- [github.com/charliermarsh/ruff-pre-commit: v0.0.237 → v0.0.241](https://github.com/charliermarsh/ruff-pre-commit/compare/v0.0.237...v0.0.241)
