+++ .pre-commit-config.yaml
@@ -1,12 +1,12 @@
 files: 'clustergram\/'
 repos:
   - repo: https://github.com/psf/black
-    rev: 22.12.0
+    rev: 23.1.0
     hooks:
       - id: black
         language_version: python3
   - repo: https://github.com/charliermarsh/ruff-pre-commit
-    rev: "v0.0.237"
+    rev: "v0.0.241"
     hooks:
       - id: ruff
 
