+++ .pre-commit-config.yaml
@@ -6,7 +6,7 @@ repos:
       - id: black
         language_version: python3
   - repo: https://github.com/charliermarsh/ruff-pre-commit
-    rev: "v0.0.241"
+    rev: "v0.0.246"
     hooks:
       - id: ruff
 
