### 2023-02-20 8d47c50
[pre-commit.ci] pre-commit autoupdate (#46)
updates:
- [github.com/charliermarsh/ruff-pre-commit: v0.0.246 → v0.0.248](https://github.com/charliermarsh/ruff-pre-commit/compare/v0.0.246...v0.0.248)

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
