### 2023-02-27 b3c9338
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/charliermarsh/ruff-pre-commit: v0.0.248 → v0.0.252](https://github.com/charliermarsh/ruff-pre-commit/compare/v0.0.248...v0.0.252)
