### 2023-03-31 #5240 @gagolews
@martinfleis I'm not going to be too stubborn, the package itself is quite nice. Let's see what others say about it...
