### 2023-05-08 c598d29
[pre-commit.ci] pre-commit autoupdate (#52)
updates:
- [github.com/charliermarsh/ruff-pre-commit: v0.0.262 → v0.0.265](https://github.com/charliermarsh/ruff-pre-commit/compare/v0.0.262...v0.0.265)

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
