### 2023-05-22 321216c
[pre-commit.ci] pre-commit autoupdate (#53)
updates:
- [github.com/charliermarsh/ruff-pre-commit: v0.0.265 → v0.0.269](https://github.com/charliermarsh/ruff-pre-commit/compare/v0.0.265...v0.0.269)

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
