### 2023-06-12 #5240 @martinfleis
> The author claims to embrace scikit-learn’s API style, however that is unfortunately not entirely achieved in my view.

Thanks for the comment! I'll have a deeper dive to see if the API can follow the style more closely. I suppose the issue will be in the difference between sklearn's estimators producing clustering labels for a fixed `k` while clustergram requires a range of results but let me see what can be done here.

>  is explicitly referencing scikit-learn’s API style, I would suggest to cite Buitinck et al.

Will do, thanks!

>  I would recommend that the ticks on the x-axis on all clustergram diagrams are ensured to be natural numbers

Good point. Will fix.

> There are no instructions on how to contribute to the package.

Contributing guide is available in the documentation: https://clustergram.readthedocs.io/en/stable/contributing.html I can eventually copy its content to CONTRIBNUTING.md and store it at the root of the repo but I didn't think the duplication is necessary here.

Thanks you for your comment!
