### 2023-06-12 martinfleis/clustergram#56 @martinfleis
Ensure x ticks are labeled with integers
From https://github.com/openjournals/joss-reviews/issues/5240#issuecomment-1586355157

> As a bit of a minor point, however since plotting is the core function of this package I think it warrants mentioning, I would recommend that the ticks on the x-axis on all clustergram diagrams are ensured to be natural numbers since there are no fractions of numbers of clusters.
