### 2023-07-18 be5dfe7
[pre-commit.ci] pre-commit autoupdate (#54)
updates:
- [github.com/psf/black: 23.3.0 → 23.7.0](https://github.com/psf/black/compare/23.3.0...23.7.0)
- [github.com/astral-sh/ruff-pre-commit: v0.0.272 → v0.0.278](https://github.com/astral-sh/ruff-pre-commit/compare/v0.0.272...v0.0.278)

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
