+++ .pre-commit-config.yaml
@@ -1,12 +1,12 @@
 files: 'clustergram\/'
 repos:
   - repo: https://github.com/psf/black
-    rev: 23.3.0
+    rev: 23.7.0
     hooks:
       - id: black
         language_version: python3
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: "v0.0.272"
+    rev: "v0.0.278"
     hooks:
       - id: ruff
 
