### 2023-07-18 d83c851
TST: adapt tests to scikit-learn 1.3.0 changes
