+++ clustergram/test_clustergram.py
@@ -202,17 +202,16 @@ def test_bic():
             306.6669136240255,
         ]
     else:
-        expected = (
-            [
-                1226.7924019554766,
-                948.6374834781362,
-                800.1788609508928,
-                687.5987056807201,
-                497.2770114251739,
-                402.1340827435864,
-                306.6669136240255,
-            ],
-        )
+        expected = [
+            1226.7924019554766,
+            948.6374834781362,
+            800.1788609508928,
+            687.5987056807201,
+            497.2770114251739,
+            402.1340827435864,
+            306.6669136240255,
+        ]
+
     expected = pd.Series(
         expected,
         index=range(1, 8),
