### 2023-07-18 3cb8147
TST: adapt tests to scikit-learn 1.3.0 changes (#57)
