### 2023-07-18 683c148
ENH: ensure that x ticks are only integers
