### 2023-07-18 fcfd9bb
conditionally + ruff
