+++ clustergram/test_clustergram.py
@@ -1,12 +1,12 @@
-from distutils.version import Version
+import bokeh
 import numpy as np
 import pandas as pd
 import pytest
+import sklearn
 from bokeh.embed import json_item
+from packaging.version import Version
 from pandas.testing import assert_series_equal
 from sklearn.datasets import make_blobs
-import sklearn
-from packaging.version import Version
 
 try:
     import cudf
@@ -20,6 +20,7 @@ except (ImportError, ModuleNotFoundError):
 from clustergram import Clustergram
 
 SKLEARN_GE_130 = Version(sklearn.__version__) >= Version("1.3.0")
+BOKEH_GE_300 = Version(bokeh.__version__) >= Version("3.0.0")
 
 n_samples = 100
 n_features = 2
@@ -61,7 +62,7 @@ def test_sklearn_kmeans():
         for x in range(1, 8)
     ]
 
-    ax = clustergram.plot(pca_kwargs=dict(random_state=random_state))
+    ax = clustergram.plot(pca_kwargs={"random_state": random_state})
     assert len(ax.get_children()) == 46
 
     assert clustergram.plot_data.empty
@@ -115,7 +116,7 @@ def test_sklearn_minibatchkmeans():
         for x in range(1, 8)
     ]
 
-    ax = clustergram.plot(pca_kwargs=dict(random_state=random_state))
+    ax = clustergram.plot(pca_kwargs={"random_state": random_state})
     assert len(ax.get_children()) == 46 if SKLEARN_GE_130 else 45
 
     assert clustergram.plot_data.empty
@@ -166,7 +167,7 @@ def test_sklearn_gmm():
         for x in range(1, 8)
     ]
 
-    ax = clustergram.plot(pca_kwargs=dict(random_state=random_state))
+    ax = clustergram.plot(pca_kwargs={"random_state": random_state})
     assert len(ax.get_children()) == 44
 
     assert clustergram.plot_data.empty
@@ -274,7 +275,7 @@ def test_cuml_kmeans():
         for x in range(1, 8)
     ]
 
-    ax = clustergram.plot(pca_kwargs=dict(random_state=random_state))
+    ax = clustergram.plot(pca_kwargs={"random_state": random_state})
     assert len(ax.get_children()) == 44
 
     assert clustergram.plot_data.empty
@@ -313,7 +314,7 @@ def test_cuml_kmeans():
         for x in range(1, 8)
     ]
 
-    ax = clustergram.plot(pca_kwargs=dict(random_state=random_state))
+    ax = clustergram.plot(pca_kwargs={"random_state": random_state})
     assert len(ax.get_children()) == 44
 
     assert clustergram.plot_data.empty
@@ -351,7 +352,7 @@ def test_hierarchical():
         for x in range(1, 8)
     ]
 
-    ax = clustergram.plot(pca_kwargs=dict(random_state=random_state))
+    ax = clustergram.plot(pca_kwargs={"random_state": random_state})
     assert len(ax.get_children()) == 44
 
     assert clustergram.plot_data.empty
@@ -684,7 +685,7 @@ def test_from_data_mean():
     labels = pd.DataFrame({1: [0, 0, 0], 2: [0, 0, 1], 3: [0, 2, 1]})
     clustergram = Clustergram.from_data(data, labels)
 
-    ax = clustergram.plot(pca_kwargs=dict(random_state=random_state))
+    ax = clustergram.plot(pca_kwargs={"random_state": random_state})
     assert len(ax.get_children()) == 18
 
     assert clustergram.plot_data.empty
@@ -704,7 +705,7 @@ def test_from_data_median():
     labels = pd.DataFrame({1: [0, 0, 0], 2: [0, 0, 1], 3: [0, 2, 1]})
[…]
