### 2023-07-18 a4123c6
move checks from init to fit
