### 2023-07-18 4609629
return self and ensure each fit call is fresh
