+++ clustergram/test_clustergram.py
@@ -77,6 +77,9 @@ def test_sklearn_kmeans():
         1.4398916223315354, rel=1e-4
     )
 
+    assert isinstance(clustergram.labels_, pd.DataFrame)
+    assert isinstance(clustergram.cluster_centers_, dict)
+
 
 def test_sklearn_minibatchkmeans():
     clustergram = Clustergram(
@@ -131,6 +134,9 @@ def test_sklearn_minibatchkmeans():
         1.4441508195691932 if SKLEARN_GE_130 else 1.477158426841248, rel=1e-4
     )
 
+    assert isinstance(clustergram.labels_, pd.DataFrame)
+    assert isinstance(clustergram.cluster_centers_, dict)
+
 
 def test_sklearn_gmm():
     clustergram = Clustergram(
@@ -182,6 +188,9 @@ def test_sklearn_gmm():
         1.4636749962003583 if SKLEARN_GE_130 else 1.3321040444661392, rel=1e-4
     )
 
+    assert isinstance(clustergram.labels_, pd.DataFrame)
+    assert isinstance(clustergram.cluster_centers_, dict)
+
 
 def test_bic():
     clustergram = Clustergram(
@@ -220,6 +229,8 @@ def test_bic():
     )
 
     assert_series_equal(expected, clustergram.bic, rtol=1e-6)
+
+    assert isinstance(clustergram.bic_, pd.Series)
 
     clustergram = Clustergram(
         range(1, 8),
@@ -367,6 +378,8 @@ def test_hierarchical():
         1.4398916223315354, rel=1e-4
     )
 
+    assert isinstance(clustergram.linkage_, np.ndarray)
+
 
 def test_hierarchical_array():
     clustergram = Clustergram(method="hierarchical", k_range=range(1, 10))
@@ -437,6 +450,8 @@ def test_silhouette_score():
             name="silhouette_score",
         ),
     )
+
+    assert isinstance(clustergram.silhouette_, pd.Series)
 
 
 @pytest.mark.skipif(
@@ -522,6 +537,8 @@ def test_calinski_harabasz_score():
             name="calinski_harabasz_score",
         ),
     )
+
+    assert isinstance(clustergram.calinski_harabasz_, pd.Series)
 
 
 @pytest.mark.skipif(
@@ -619,6 +636,8 @@ def test_davies_bouldin_score():
             name="davies_bouldin_score",
         ),
     )
+
+    assert isinstance(clustergram.davies_bouldin_, pd.Series)
 
 
 @pytest.mark.skipif(
@@ -901,13 +920,6 @@ def test_custom_pca_cuml():
 
 
 def test_plot_integer_ticks():
-    device_data, device_labels = make_blobs(
-        n_samples=n_samples,
-        n_features=n_features,
-        centers=n_clusters,
-        random_state=random_state,
-        cluster_std=0.1,
-    )
     clustergram = Clustergram(
         range(1, 5),
         random_state=random_state,
