### 2023-07-18 1475a8a
API: follow sklearn api conventions (#59)
* move checks from init to fit

* make fit signature follow convention

* return self and ensure each fit call is fresh

* estimated attributes with underscores
