+++ clustergram/clustergram.py
@@ -68,14 +68,14 @@ class Clustergram:
 
     Attributes
     ----------
-    labels : DataFrame
-        DataFrame with cluster labels for each iteration.
-    cluster_centers : dict
-        Dictionary with cluster centers for each iteration.
-    linkage : scipy.cluster.hierarchy.linkage
-        Linkage object for hierarchical methods.
-    bic : Series
-        Bayesian Information Criterion for each iteration for Gaussian Mixture Model.
+    labels_ : DataFrame
+        DataFrame with cluster labels for each option.
+    cluster_centers_ : dict
+        Dictionary with cluster centers for each option.
+    linkage_ : numpy.ndarray
+        The hierarchical clustering encoded as a linkage matrix.
+    bic_ : Series
+        Bayesian Information Criterion for each option for Gaussian Mixture Model.
         Stored only if ``method='gmm'`` and ``bic=True``
 
 
@@ -113,84 +113,29 @@ class Clustergram:
         **kwargs,
     ):
         self.k_range = k_range
-
-        # cleanup after API change
-        kwargs.pop("pca_weighted", None)
-        kwargs.pop("pca_kwargs", None)
-
-        self.store_bic = kwargs.pop("bic", False)
-
-        if backend is None:
-            backend = "scipy" if method == "hierarchical" else "sklearn"
-
-        allowed_backends = ["sklearn", "cuML", "scipy"]
-        if backend not in allowed_backends:
-            raise ValueError(
-                f'"{backend}" is not a supported backend. '
-                f"Use one of {allowed_backends}."
-            )
-        else:
-            self.backend = backend
-
-        supported = ["kmeans", "gmm", "minibatchkmeans", "hierarchical"]
-        class_methods = ["from_centers", "from_data"]
-        if method not in supported and method not in class_methods:
-            raise ValueError(
-                f"'{method}' is not a supported method. "
-                f"Only {supported} are supported now."
-            )
-        else:
-            self.method = method
-
-        if self.k_range is None and self.method != "hierarchical":
-            raise ValueError(f"'k_range' is mandatory for '{self.method}' method.")
-
-        if (
-            (self.backend == "cuML" and self.method != "kmeans")
-            or (self.backend == "scipy" and self.method != "hierarchical")
-            or (self.backend == "sklearn" and self.method == "hierarchical")
-        ):
-            raise ValueError(
-                f"'{self.method}' method is not implemented "
-                f"for '{self.backend}' backend. Use supported combination."
-            )
-
-        self.engine_kwargs = kwargs
+        self.backend = backend
+        self.method = method
         self.verbose = verbose
-        self.link_pca = defaultdict(dict)
-
-        if self.backend in ["sklearn", "scipy"]:
-            self.plot_data = pd.DataFrame()
-            self.plot_data_pca = defaultdict(pd.DataFrame)
-
-        else:
-            try:
-                import cudf
-            except (ImportError, ModuleNotFoundError) as e:
-                raise ImportError(
-                    "cuML, cuDF and cupy packages are required to use `cuML` backend."
-                ) from e
-
-            self.plot_data = cudf.DataFrame()
-            self.plot_data_pca = defaultdict(cudf.DataFrame)
-
-        self._n_pca = 0
+        self.kwargs = kwargs
+        self._backend = backend
 
     def __repr__(self):
         return (
-            f"Clustergram(k_range={self.k_range}, backend='{self.backend}', "
-            f"method='{self.method}', kwargs={self.engine_kwargs})"
+            f"Clustergram(k_range={self.k_range}, backend='{self._backend}', "
[…]
