### 2023-07-18 025cb51
ENH: format time in verbose mode
