+++ doc/contributing.md
@@ -1,173 +1,2 @@
-
-# Contributing to Clustergram
-
-First off, thanks for taking the time to contribute! ❤️
-
-All types of contributions are encouraged and valued. See this page for different ways
-to help and details about how this project handles them. Please make sure to read the
-relevant section before making your contribution. It will make it a lot easier for us
-maintainers and smooth out the experience for all involved. The community looks forward
-to your contributions. 🎉
-
-> And if you like the project, but just don't have time to contribute, that's fine.
-> There are other easy ways to support the project and show your appreciation, which we
-> would also be very happy about:
->
-> - Star the project
-> - Tweet about it
-> - Refer this project in your project's readme
-> - Mention the project at local meetups and tell your friends/colleagues
-
-## I Have a Question
-
-> If you want to ask a question, we assume that you have read the available
-> [Documentation](https://clustergram.readthedocs.io/).
-
-Before you ask a question, it is best to search for existing
-[Issues](https://github.com/martinfleis/clustergram/issues) that might help you. In case
-you have found a suitable issue and still need clarification, you can write your
-question in this issue. It is also advisable to search the internet for answers first.
-
-If you then still feel the need to ask a question and need clarification, we recommend
-the following:
-
-- Open an [Issue](https://github.com/martinfleis/clustergram/issues/new).
-- Provide as much context as you can about what you're running into.
-- Provide project and platform versions (`python`, `scikit-learn`, `bokeh`, etc.),
-  depending on what seems relevant.
-
-We will then take care of the issue as soon as possible.
-
-## I Want To Contribute
-
-### Reporting Bugs
-
-#### Before Submitting a Bug Report
-
-A good bug report shouldn't leave others needing to chase you up for more information.
-Therefore, we ask you to investigate carefully, collect information and describe the
-issue in detail in your report. Please complete the following steps in advance to help
-us fix any potential bug as fast as possible.
-
-- Make sure that you are using the latest version.
-- Determine if your bug is really a bug and not an error on your side e.g. using
-  incompatible environment components/versions (Make sure that you have read the
-  [documentation](https://clustergram.readthedocs.io/).
-- To see if other users have experienced (and potentially already solved) the same issue
-  you are having, check if there is not already a bug report existing for your bug or
-  error in the [bug tracker](https://github.com/martinfleis/clustergram/issues).
-- Also make sure to search the internet (including Stack Overflow) to see if users
-  outside of the GitHub community have discussed the issue.
-- Collect information about the bug:
-- Stack trace (Traceback)
-- OS, Platform and Version (Windows, Linux, macOS, x86, ARM)
-- Version of Python and relevant dependencies.
-- Possibly your input and the output
-- Can you reliably reproduce the issue? And can you also reproduce it with older
-  versions?
-
-#### How Do I Submit a Good Bug Report?
-
-We use GitHub issues to track bugs and errors. If you run into an issue with the
-project:
-
-- Open an [Issue](https://github.com/martinfleis/clustergram/issues/new). (Since we
-  can't be sure at this point whether it is a bug or not, we ask you not to talk about a
-  bug yet and not to label the issue.)
-- Explain the behavior you would expect and the actual behavior.
-- Please provide as much context as possible and describe the *reproduction steps* that
-  someone else can follow to recreate the issue on their own. This usually includes your
-  code. For good bug reports you should isolate the problem and create a reduced test
-  case.
-- Provide the information you collected in the previous section.
-
-Once it's filed:
-
-- The project team will label the issue accordingly.
-- A team member will try to reproduce the issue with your provided steps. If there are
-  no reproduction steps or no obvious way to reproduce the issue, the team will ask you
[…]
