+++ paper/paper.bib
@@ -211,3 +211,38 @@
   langid     = {english},
   keywords   = {Ageing,Area classification,Geodemographics,Population}
 }
+
+@inproceedings{sklearn_api,
+  author    = {Lars Buitinck and Gilles Louppe and Mathieu Blondel and
+               Fabian Pedregosa and Andreas Mueller and Olivier Grisel and
+               Vlad Niculae and Peter Prettenhofer and Alexandre Gramfort
+               and Jaques Grobler and Robert Layton and Jake VanderPlas and
+               Arnaud Joly and Brian Holt and Ga{\"{e}}l Varoquaux},
+  title     = {{API} design for machine learning software: experiences from the scikit-learn
+               project},
+  booktitle = {ECML PKDD Workshop: Languages for Data Mining and Machine Learning},
+  year      = {2013},
+  pages     = {108--122}
+}
+
+@manual{ecotone,
+  title  = {EcotoneFinder: Characterising and Locating Ecotones and Communities},
+  author = {Antoine Bagnaro},
+  year   = {2021},
+  note   = {R package version 0.2.3},
+  url    = {https://cran.r-project.org/web/packages/EcotoneFinder/index.html}
+}
+
+@article{GAGOLEWSKI2021620,
+title = {Are cluster validity measures (in) valid?},
+journal = {Information Sciences},
+volume = {581},
+pages = {620-636},
+year = {2021},
+issn = {0020-0255},
+doi = {https://doi.org/10.1016/j.ins.2021.10.004},
+url = {https://www.sciencedirect.com/science/article/pii/S0020025521010082},
+author = {Marek Gagolewski and Maciej Bartoszuk and Anna Cena},
+keywords = {Clustering methodology, Cluster validity index, Dunn index, Nearest neighbours (NNs), Ordered weighted averaging (OWA) operator, No free lunch},
+abstract = {Internal cluster validity measures (such as the Calinski–Harabasz, Dunn, or Davies–Bouldin indices) are frequently used for selecting the appropriate number of partitions a dataset should be split into. In this paper we consider what happens if we treat such indices as objective functions in unsupervised learning activities. Is the optimal grouping with regards to, say, the Silhouette index really meaningful? It turns out that many cluster (in)validity indices promote clusterings that match expert knowledge quite poorly. We also introduce a new, well-performing variant of the Dunn index that is built upon OWA operators and the near-neighbour graph so that subspaces of higher density, regardless of their shapes, can be separated from each other better.}
+}
