### 2023-07-19 #5240 @martinfleis
I have finished a review of the package and the paper based on the comments by @csadorf and @gagolews above. Below are responses to points that have not been discussed in the individual comments in above.

From the @gagolews's review (plus see https://github.com/openjournals/joss-reviews/issues/5240#issuecomment-1490353259):

> It seems that there is an implementation of the clustergrams in the EcotoneFinder package for R, see <https://cran.r-project.org/web/packages/EcotoneFinder/index.html >, this probably should be mentioned as well.

EcotoneFinder is now mentioned in the paper, I wasn't previously aware of the implementation.

> Other remarks: [...]

All the comments related to the paper have been resolved in https://github.com/martinfleis/clustergram/commit/52676ab7966f0d3a183e48b1a93af79d6ea2e55b

From the @csadorf's review:

> The author claims to embrace scikit-learn’s API style, however that is unfortunately not entirely achieved in my view. While the main class does provide a estimator-like interface, accepting hyper parameters as constructor arguments and offering a fit() function, other parts of the [scikit-learn API guide](https://scikit-learn.org/stable/developers/develop.html) are ignored, such as avoiding the modification of hyper parameters within the constructor, storing fitted parameters with an underscore suffix, and allowing for easy composability. Other plotting-classes within the scikit-learn package typically implement “from_estimator” and “from_predictions” class methods (e.g. ConfusionMatrixDisplay, PredictionErrorDisplay, etc.). In this way it is easy to use the class with any class that would provide clustering data. The current implementation of the class provides from_data() and from_centers() methods, which achieve similar, but not directly compatible behavior.

The API have been adjusted in https://github.com/martinfleis/clustergram/pull/59. See the PR for the details.

> Since the author is explicitly referencing scikit-learn’s API style, I would suggest to cite Buitinck et al. ([arXiv:1309.0238](https://arxiv.org/abs/1309.0238) [cs.LG]) as recommended by the [scikit-learn's citation guide](https://scikit-learn.org/stable/about.html#citing-scikit-learn).

Included now.

> As a bit of a minor point, however since plotting is the core function of this package I think it warrants mentioning, I would recommend that the ticks on the x-axis on all clustergram diagrams are ensured to be natural numbers since there are no fractions of numbers of clusters.

Both plotting methods (matplotlib, bokeh), are now limited to integer ticks only.

> There are no instructions on how to contribute to the package.

As mentioned in https://github.com/openjournals/joss-reviews/issues/5240#issuecomment-1587065252, the contributing guide was part of the documentation. It is now also available as `CONTRIBUTING.md` in the root of the repo.

---
@csoneson  I believe that all comments are now either resolved or discussed in my earlier replies.
