### 2023-07-24 #5240 @gagolews
I deem the revised version acceptable. Well done.
