### 2023-08-01 #5240 @martinfleis
Thank you @csadorf and @gagolews!

@csoneson I have released the updated version as `v0.8.0`. Its zenodo DOI is `10.5281/zenodo.8202396`.
