+++ paper/paper.bib
@@ -185,10 +185,17 @@
   keywords = {Arctic lakes,Forel-Ulle index,Lakes classification,Satellite monitoring,Svalbard}
 }
 @article{raschka2020machine,
-  title   = {Machine Learning in Python: Main developments and technology trends in data science, machine learning, and artificial intelligence},
-  author  = {Raschka, Sebastian and Patterson, Joshua and Nolet, Corey},
-  journal = {arXiv preprint arXiv:2002.04803},
-  year    = {2020}
+  author         = {Raschka, Sebastian and Patterson, Joshua and Nolet, Corey},
+  title          = {Machine Learning in Python: Main Developments and Technology Trends in Data Science, Machine Learning, and Artificial Intelligence},
+  journal        = {Information},
+  volume         = {11},
+  year           = {2020},
+  number         = {4},
+  article-number = {193},
+  url            = {https://www.mdpi.com/2078-2489/11/4/193},
+  issn           = {2078-2489},
+  abstract       = {Smarter applications are making better use of the insights gleaned from data, having an impact on every industry and research discipline. At the core of this revolution lies the tools and the methods that are driving it, from processing the massive piles of data generated each day to learning from and taking useful action. Deep neural networks, along with advancements in classical machine learning and scalable general-purpose graphics processing unit (GPU) computing, have become critical components of artificial intelligence, enabling many of these astounding breakthroughs and lowering the barrier to adoption. Python continues to be the most preferred language for scientific computing, data science, and machine learning, boosting both performance and productivity by enabling the use of low-level libraries and clean high-level APIs. This survey offers insight into the field of machine learning with Python, taking a tour through important topics to identify some of the core hardware and software paradigms that have enabled it. We cover widely-used libraries and concepts, collected together for holistic comparison, with the goal of educating the reader and driving the field of Python machine learning forward.},
+  doi            = {10.3390/info11040193}
 }
 @manual{palmerpenguins,
   title  = {palmerpenguins: Palmer Archipelago (Antarctica) penguin data},
@@ -234,15 +241,15 @@
 }
 
 @article{GAGOLEWSKI2021620,
-title = {Are cluster validity measures (in) valid?},
-journal = {Information Sciences},
-volume = {581},
-pages = {620-636},
-year = {2021},
-issn = {0020-0255},
-doi = {https://doi.org/10.1016/j.ins.2021.10.004},
-url = {https://www.sciencedirect.com/science/article/pii/S0020025521010082},
-author = {Marek Gagolewski and Maciej Bartoszuk and Anna Cena},
-keywords = {Clustering methodology, Cluster validity index, Dunn index, Nearest neighbours (NNs), Ordered weighted averaging (OWA) operator, No free lunch},
-abstract = {Internal cluster validity measures (such as the Calinski–Harabasz, Dunn, or Davies–Bouldin indices) are frequently used for selecting the appropriate number of partitions a dataset should be split into. In this paper we consider what happens if we treat such indices as objective functions in unsupervised learning activities. Is the optimal grouping with regards to, say, the Silhouette index really meaningful? It turns out that many cluster (in)validity indices promote clusterings that match expert knowledge quite poorly. We also introduce a new, well-performing variant of the Dunn index that is built upon OWA operators and the near-neighbour graph so that subspaces of higher density, regardless of their shapes, can be separated from each other better.}
+  title    = {Are cluster validity measures (in) valid?},
+  journal  = {Information Sciences},
+  volume   = {581},
+  pages    = {620-636},
+  year     = {2021},
+  issn     = {0020-0255},
+  doi      = {10.1016/j.ins.2021.10.004},
+  url      = {https://www.sciencedirect.com/science/article/pii/S0020025521010082},
[…]
