### 2023-08-04 #5240 @csoneson
👋🏻 @csadorf - as I noticed you reacted with 👍🏻 to the message below; could you also tick the remaining boxes in https://github.com/openjournals/joss-reviews/issues/5240#issuecomment-1559574286, to make sure everything is documented? Of course, let us know if you need more time! Thanks 🙂

> Thank you @csadorf - could you please also tick the remaining boxes in the checklist above, for completeness?
