### 2023-08-30 #5240 @martinfleis
@csadorf Thanks! I have updated Zenodo as requested. DOI of the latest fixed versions is `10.5281/zenodo.8202396`.

> one reference (Pedregosa et al) is missing a DOI but I don't think it actually has one

That was my understanding as well.

All tasks from the author checklist above (checks etc) have also been done.
