### 2023-08-30 #5240 @csoneson
@martinfleis It seems the editorial bot is not happy with the two short DOIs (although the links in the bibliography go to the right place). There are long forms for both these DOIs - would you mind switching to those?
