### 2023-08-30 #5240 @csoneson
@openjournals/dev - I'm not sure whether this is an expected error, or if these short DOIs should be accepted ☝🏻. The links in the bibliography are  https://doi.org/ghh97z and  https://doi.org/ggj45f, which do point to the right places.
