+++ paper/paper.bib
@@ -15,7 +15,7 @@
   journal = {Nature Methods},
   volume  = {17},
   pages   = {261--272},
-  doi     = {10/ggj45f},
+  doi     = {10.1038/s41592-019-0686-2},
   adsurl  = {https://rdcu.be/b08Wh}
 }
 
@@ -115,7 +115,7 @@
   number    = {4},
   pages     = {391--402},
   publisher = {{SAGE Publications Sage CA: Los Angeles, CA}},
-  doi       = {10/ghh97z}
+  doi       = {10.1177/1536867X0200200405}
 }
 
 @article{schonlau2004Visualizing,
